Fix all Kotlin compile warnings
Zero warnings across compileKotlinMingwX64, compileTestKotlinMingwX64,
mingwX64Test, and the cli/gui release link tasks.

Five categories addressed:

1. Unnecessary safe calls on smart-cast non-null receivers
   (ExtractionTui.kt, ImGuiRenderer.kt). Inside `when (stage) { … }`
   branches that match a specific ExtractionStage value, the compiler
   smart-casts `progress` to non-null because the stage came from
   `progress?.stage` and matched a non-null value. Replaced
   `progress?.field` with `progress.field` and dropped the now-unreachable
   `?: 0` defaults. Hoisted `state.progress` into a local in
   ImGuiRenderer.drawExtraction so the smart cast survives across uses.

2. Redundant `.toLong()` on `statBuf.st_mtime` in WindowsArchiveRepository
   and WindowsFileSystemRepository — st_mtime is already Long on x64 K/N.

3. Mock parameter name mismatches in ApplicationViewModelTest and
   ExtractionViewModelTest. Their inline FileSystemRepository overrides
   used `path`/`parts` while the interface declares `filePath`/`components`
   for `getParentDirectory`, `joinPath`, and `moveToTrash`. Named-argument
   call sites that referred to the supertype names would fail; renamed the
   override params to match the interface.

4. ExperimentalCoroutinesApi opt-in in ExtractionViewModelTest. Added
   `@file:OptIn(kotlinx.coroutines.ExperimentalCoroutinesApi::class)` so
   the use of `runTest`/`advanceUntilIdle`/`TestScope` etc. doesn't warn
   on every call.

Tests still pass; release linker still emits clean qunzip.exe and
QuickUnzip.exe.

+++ src/mingwX64Main/kotlin/qunzip/platform/WindowsArchiveRepository.kt
@@ -209,7 +209,7 @@ class WindowsArchiveRepository(
         FileInfo(
             path = path,
             size = statBuf.st_size.toLong(),
-            lastModified = kotlinx.datetime.Instant.fromEpochSeconds(statBuf.st_mtime.toLong()),
+            lastModified = kotlinx.datetime.Instant.fromEpochSeconds(statBuf.st_mtime),
             isReadable = true,
             isDirectory = (statBuf.st_mode.toInt() and S_IFDIR) != 0
         )
